feature: refactor query execution to make future changes easier
This commit removes the SQLQuery class and simplifies the QueryResult class. It also moves the execute and execute_with_logging methods from the Query model to a utils function, allowing for a better seperation of concerns. Some unused functionality was also removed such as column summaries
